Moved components to subfolders and enabled conf tests
