+++ tests/conformance/common.go
@@ -50,7 +50,8 @@ import (
 	b_redis "github.com/dapr/components-contrib/bindings/redis"
 	p_snssqs "github.com/dapr/components-contrib/pubsub/aws/snssqs"
 	p_eventhubs "github.com/dapr/components-contrib/pubsub/azure/eventhubs"
-	p_servicebustopics "github.com/dapr/components-contrib/pubsub/azure/servicebus.topics"
+	p_servicebusqueues "github.com/dapr/components-contrib/pubsub/azure/servicebus/queues"
+	p_servicebustopics "github.com/dapr/components-contrib/pubsub/azure/servicebus/topics"
 	p_hazelcast "github.com/dapr/components-contrib/pubsub/hazelcast"
 	p_inmemory "github.com/dapr/components-contrib/pubsub/in-memory"
 	p_jetstream "github.com/dapr/components-contrib/pubsub/jetstream"
@@ -361,8 +362,10 @@ func loadPubSub(tc TestComponent) pubsub.PubSub {
 		pubsub = p_redis.NewRedisStreams(testLogger)
 	case eventhubs:
 		pubsub = p_eventhubs.NewAzureEventHubs(testLogger)
-	case "azure.servicebus":
+	case "azure.servicebus.topics":
 		pubsub = p_servicebustopics.NewAzureServiceBusTopics(testLogger)
+	case "azure.servicebus.queues":
+		pubsub = p_servicebusqueues.NewAzureServiceBusQueues(testLogger)
 	case "natsstreaming":
 		pubsub = p_natsstreaming.NewNATSStreamingPubSub(testLogger)
 	case "jetstream":
